LayoutTests/platform/ipad/fast/css/webkit-overflow-scrolling-parsing-modern-compatibility-mode-expected.txt

This tests checks parsing of the '-webkit-overflow-scrolling' property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-overflow-scrolling: auto", "-webkit-overflow-scrolling") is ""
PASS test("-webkit-overflow-scrolling: banana", "-webkit-overflow-scrolling") is ""
PASS test("-webkit-overflow-scrolling: touch", "-webkit-overflow-scrolling") is ""
PASS successfullyParsed is true

TEST COMPLETE
